requirements.txt
Django==4.2.5
django-environ==0.11.2
psycopg==3.1.12
gunicorn==23.0.0
whitenoise==6.7.0
Pillow==10.4.0
django-cloudinary-storage==0.3.0
python-magic==0.4.27
django-basicauth==0.5.3